requirements.txt
# Pipeline dependencies (used by GitHub Actions scheduler)
groq
google-generativeai
python-dotenv
pandas
google-play-scraper
app-store-scraper
emoji
google-api-python-client
google-auth
mcp

# Web app dependencies (used locally only)
fastapi
uvicorn
streamlit
pydantic
email-validator
resend
